Add collect_set Spark aggregation function
