+++ velox/functions/sparksql/aggregates/CMakeLists.txt
@@ -18,6 +18,7 @@ velox_add_library(
   BloomFilterAggAggregate.cpp
   CentralMomentsAggregate.cpp
   CollectListAggregate.cpp
+  CollectSetAggregate.cpp
   FirstLastAggregate.cpp
   MinMaxAggregate.cpp
   MinMaxByAggregate.cpp
