+++ velox/functions/sparksql/aggregates/tests/CMakeLists.txt
@@ -19,6 +19,7 @@ add_executable(
   BloomFilterAggAggregateTest.cpp
   CentralMomentsAggregationTest.cpp
   CollectListAggregateTest.cpp
+  CollectSetAggregateTest.cpp
   FirstAggregateTest.cpp
   LastAggregateTest.cpp
   Main.cpp
@@ -35,6 +36,7 @@ target_link_libraries(
   velox_exec_test_lib
   velox_vector_test_lib
   velox_functions_aggregates_test_lib
+  velox_functions_spark
   velox_functions_spark_aggregates
   velox_hive_connector
   velox_vector_fuzzer
